added frontend react to attendance selector
